requirement.txt
Django==1.4.1
Pillow==1.7.7
argparse==1.2.1
django-appconf==0.5
django-imagekit==2.0.2
wsgiref==0.1.2
